refactor and bugfixing
