+++ backend/src/testData/logTablesData.ts
@@ -1,4 +1,5 @@
-import db from "./testDatabase.js";
+// import db from "./testDatabase.js";
+import db from "../database.js";
 
 // Log all rows from the 'transfers' table
 db.all("SELECT * FROM transfers", (err, rows) => {
